Fyyur/requirements.txt
alembic==1.8.0
Babel==2.10.3
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.2.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
